app/pyproject.toml
[tool.poetry]
name = "broken-toy-shop"
version = "0.1.0"
description = ""
authors = ["Sergei Konik <s.konik.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "broken_toy_shop"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^21.2.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
absolufy-imports = "^0.3.1"
autoflake = "^2.2.0"
ruff = "^0.0.280"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.git-limiter]
max-changed-files = 6
max-insertions = 300
max-deletions = 200

[tool.isort]
line_length = 100
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_output = 3
known_django = "django"
profile = "black"
src_paths = "."
lines_after_imports = 2



[tool.ruff]
target-version = "py311"
fix = true
line-length = 100

select = [
    "F",   # pyflakes
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "C90", # mccabe
    "N",   # pep8-naming
    "DJ",  # flake8-django
    "C4",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "BLE", # flake8-blind-except
    "T20", # flake8-print
    "Q",   # flake8-quotes
]
exclude = [
    "eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".tox",
    ".venv",
    ".pyc",
    ".pyi",
    "_build",
    "buck-out",
    "build",
    "dist",
    "migrations",
]

[tool.ruff.mccabe]
max-complexity = 10


[tool.black]
line-length = 100
target-version = ['py311']
exclude = '''
(
  \.eggs
  |\.git
  |\.hg
  |\.mypy_cache
  |\.nox
  |\.tox
  |\.venv
  |\.pyc
  |\.pyi
  |_build
  |buck-out
  |build
  |dist
)
'''